Setup suite in test package init
